Add configurable default landing pages and handle relative URLs
This adds configurable default landing pages for redmine
and projects that don't have a landing page defined.

If nothing is defined as a landing page the redmine
defaults are used.

There's a small change in behaviour that makes the original
landing pages work again if they are called directly, so
no functionality is lost when using landing pages.

Additionally to be more robust to server changes allow
relatives URLs to reference modules in the same project/scope

Signed-off-by: Tobias Droste <tdroste@gmx.de>